requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
python-multipart
Pillow